Crate wasm4_sys

Constants

BLIT_1BPP

BLIT_2BPP

BLIT_FLIP_X

BLIT_FLIP_Y

BLIT_ROTATE

BUTTON_1

BUTTON_2

BUTTON_DOWN

BUTTON_LEFT

BUTTON_RIGHT

BUTTON_UP

DRAW_COLORS

FRAMEBUFFER

GAMEPAD1

GAMEPAD2

GAMEPAD3

GAMEPAD4

MOUSE_BUTTONS

MOUSE_LEFT

MOUSE_MIDDLE

MOUSE_RIGHT

MOUSE_X

MOUSE_Y

NETPLAY

PALETTE

SCREEN_SIZE

SYSTEM_FLAGS

SYSTEM_HIDE_GAMEPAD_OVERLAY

SYSTEM_PRESERVE_FRAMEBUFFER

TONE_MODE1

TONE_MODE2

TONE_MODE3

TONE_MODE4

TONE_NOISE

TONE_PAN_LEFT

TONE_PAN_RIGHT

TONE_PULSE1

TONE_PULSE2

TONE_TRIANGLE

Functions

blit

Copies pixels to the framebuffer.

blit_sub

Copies a subregion within a larger sprite atlas to the framebuffer.

diskr⚠

Reads up to size bytes from persistent storage into the pointer dest.

diskw⚠

Writes up to size bytes from the pointer src into persistent storage.

hline

Draws a horizontal line

line

Draws a line between two points.

oval

Draws an oval (or circle).

rect

Draws a rectangle.

text

Draws text using the built-in system font.

tone

Plays a sound tone.

trace

Prints a message to the debug console.

vline

Draws a vertical line